requirements.txt
telethon
pillow
matrix-nio
pyyaml
aiofiles
python-magic
lottie
cairosvg
tqdm